fix bug for create canlde in to images error
