+++ internal/bot/commands.go
@@ -1,0 +1,1 @@
+package telegrambot
