+++ internal/bot/handlers.go
@@ -1,0 +1,16 @@
+package telegrambot
+
+import tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
+
+func (b *Bot) handleUpdate(updates tgbotapi.UpdatesChannel) {
+	for update := range updates {
+		if update.Message != nil {
+			b.handleMessage(update.Message)
+		}
+	}
+}
+
+func (b *Bot) handleMessage(message *tgbotapi.Message) {
+	msg := tgbotapi.NewMessage(message.Chat.ID, message.Text)
+	b.bot.Send(msg)
+}
