+++ internal/database/postgres/candles/candles.go
@@ -5,6 +5,7 @@ import (
 	"github.com/Sanchir01/candles_backend/internal/gql/model"
 	"github.com/google/uuid"
 	"github.com/jmoiron/sqlx"
+	"github.com/lib/pq"
 	"github.com/samber/lo"
 	"time"
 )
@@ -33,15 +34,17 @@ func (s *CandlesPostgresStore) AllCandles(ctx context.Context) ([]model.Candles,
 
 	return lo.Map(candles, func(candles dbCandles, _ int) model.Candles { return model.Candles(candles) }), nil
 }
-func (s *CandlesPostgresStore) CreateCandles(ctx context.Context, categoryID uuid.UUID, title string, slug string) (uuid.UUID, error) {
+func (s *CandlesPostgresStore) CreateCandles(
+	ctx context.Context, categoryID uuid.UUID, title string, slug string,
+	images []string) (uuid.UUID, error) {
 	conn, err := s.db.Connx(ctx)
 	if err != nil {
 		return uuid.Nil, err
 	}
 	defer conn.Close()
 	var id uuid.UUID
-
-	if err := conn.GetContext(ctx, &id, "INSERT INTO candles (category_id, title, slug) VALUES ($1, $2, $3) RETURNING id", categoryID, title, slug); err != nil {
+	pqImages := pq.Array(images)
+	if err := conn.GetContext(ctx, &id, "INSERT INTO candles (category_id, title, slug, images) VALUES ($1, $2, $3, $4) RETURNING id", categoryID, title, slug, pqImages); err != nil {
 		return uuid.Nil, err
 	}
 	return id, nil
@@ -78,7 +81,7 @@ func (s *CandlesPostgresStore) CandlesById(ctx context.Context, id uuid.UUID) (*
 
 type dbCandles struct {
 	ID         uuid.UUID `db:"id"`
-	Title      string    `db:"name"`
+	Title      string    `db:"title"`
 	Slug       string    `db:"slug"`
 	CreatedAt  time.Time `db:"created_at"`
 	UpdatedAt  time.Time `db:"updated_at"`
